refactor(bpf): collapse five run_bpf2go.go wrappers into shared helper
The 5 BPF probe packages that dispatch by syscall NR (traceconnect,
tracedns, tracefs, tracebpfaudit, tracelsmenforce) each shipped a
near-identical ~55-line run_bpf2go.go that resolved repo-root, derived
-D__TARGET_ARCH_<runtime.GOARCH>, and exec'd bpf2go. The only meaningful
differences were the target name (Traceconnect, Tracedns, …) and the
.bpf.c source file.

Extract one //go:build ignore main program under
internal/bpf/bpfgen/main.go that takes <Target> <basename> as args, and
have each package's gen.go invoke it via
//go:generate go run ../bpfgen/main.go <Target> <source>.bpf.c.

Net: ~280 LOC removed, single point of edit for cflags / bpf2go version
bumps. //go:build ignore keeps the helper out of regular go build ./....
