--- internal/bpf/tracebpfaudit/run_bpf2go.go
@@ -1,53 +1,0 @@
-//go:build ignore
-
-package main
-
-import (
-	"log"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"runtime"
-)
-
-func main() {
-	pkgDir, err := os.Getwd()
-	if err != nil {
-		log.Fatal(err)
-	}
-	repoRoot := filepath.Clean(filepath.Join(pkgDir, "..", "..", ".."))
-
-	archFlag := "-D__TARGET_ARCH_x86"
-	if runtime.GOARCH == "arm64" {
-		archFlag = "-D__TARGET_ARCH_arm64"
-	}
-
-	cflags := archFlag +
-		" -O2 -g -Wall -Werror" +
-		" -I" + filepath.Join(repoRoot, "bpf") +
-		" -I/usr/include/bpf"
-	extra := []string{
-		"-I" + filepath.Join(repoRoot, "bpf"),
-		"-I/usr/include/bpf",
-	}
-
-	args := []string{
-		"run", "github.com/cilium/ebpf/cmd/bpf2go@v0.21.0",
-		"-cc", "clang",
-		"-no-strip",
-		"-target", "bpfel,bpfeb",
-		"-cflags", cflags,
-		"Tracebpfaudit",
-		filepath.Join(repoRoot, "bpf", "trace_bpf_audit.bpf.c"),
-		"--",
-	}
-	args = append(args, extra...)
-
-	cmd := exec.Command("go", args...)
-	cmd.Dir = pkgDir
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err := cmd.Run(); err != nil {
-		log.Fatal(err)
-	}
-}
