+++ internal/bpf/tracedns/gen.go
@@ -1,14 +1,9 @@
 // Package tracedns loads the trace_dns.bpf.c BPF object via cilium/ebpf
-// bpf2go-generated bindings.
-//
-// Loader pattern: indirect via run_bpf2go.go (//go:build ignore main program).
-// trace_dns.bpf.c includes trace_connect_obs.h which dispatches recvfrom by
-// architecture-specific syscall NR (NR_recvfrom). bpf2go must therefore be
-// invoked with `-D__TARGET_ARCH_<arch>` derived from `runtime.GOARCH`; the
-// run_bpf2go.go helper builds that flag string at generate time.
-//
-// See internal/bpf/traceconnect/gen.go for a fuller explanation of the two
-// loader patterns used in this repo.
+// bpf2go-generated bindings. trace_dns.bpf.c includes trace_connect_obs.h
+// which dispatches recvfrom by architecture-specific syscall NR
+// (NR_recvfrom), so bpf2go must be invoked with `-D__TARGET_ARCH_<arch>`
+// derived from `runtime.GOARCH`. See internal/bpf/bpfgen/main.go and
+// internal/bpf/traceconnect/gen.go for the shared rationale.
 package tracedns
 
-//go:generate go run ./run_bpf2go.go
+//go:generate go run ../bpfgen/main.go Tracedns trace_dns.bpf.c
